Implement contract json validation endpoint (#10)
The new endpoint is available at POST /contract/validate. It accepts
json requests with two fields: `contract` with the contract json and
`contract_hash` with the sha256 hash of the canonicalized json.

This also promotes `version` and `issuer_pubkey` to first-class fields
available on AssetFields and directly on the root json document, which
requires a matching update in the json format used by the asset git db repo.